pyproject.toml
[tool.poetry]
name = "chords"
version = "0.1.0"
description = "Song book written in Django, supporting offline viewing"
authors = ["Petr Hála <phala@sprtovo.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "==4.1.10"
django-bootstrap4 = "*"
dj-datatables-view = "*"
django-debug-toolbar = "*"
django-markdownx = ">=4.0.0b1"
django-redis = "*"
django-rq = "*"
django-sass-processor = "*"
django-simple-menu = "*"
django-weasyprint = "*"
weasyprint = ">=53.0"
Pillow = "*"
gunicorn = "*"
gevent = "*"
markdown = "*"
markdown3-newtab = "*"
netifaces = "*"
psycopg2-binary = "*"
django-compressor = "*"
libsass = "*"

[tool.poetry.group.dev.dependencies]
pylint = "*"
pylint-django = "*"
black = "*"

[tool.black]
line-length = 120

# pylint
[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.DESIGN]
min-public-methods=1
max-parents=12
disable = [
    "duplicate-code", # reports false alarms AND can't be disabled locally; pylint issue #214
    "fixme", # ignore TODOs
    "redefined-outer-name",
    "too-many-arguments",
    "too-few-public-methods"]

[tool.pylint.MAIN]
ignore-patterns= "\\d{4}_.*?.py"
load-plugins = "pylint_django"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
